requirements.txt
requests>=2.28.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.0.0
